Migrates SongSearch into separate component
