+++ src/modules/songs/components/SongSearch/index.ts
@@ -1,0 +1,1 @@
+export * from "./SongSearch";
